Generate command interactive option by default added to sergen
Sample project removed
