--- Serenity.Net.CodeGenerator.TerminalGui/Serenity.Net.CodeGenerator.TerminalGui.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Spectre.Console" Version="0.43.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Serenity.Net.CodeGenerator\Serenity.Net.CodeGenerator.csproj" />
-  </ItemGroup>
-
-</Project>
