+++ src/Serenity.Net.CodeGenerator/Serenity.Net.CodeGenerator.csproj
@@ -35,6 +35,7 @@
     <PackageReference Include="Microsoft.Build" Version="16.8.0" ExcludeAssets="runtime" />
     <PackageReference Include="Microsoft.Build.Utilities.Core" Version="16.8.0" ExcludeAssets="runtime" />
     <PackageReference Include="Microsoft.Build.Locator" Version="1.3.2" />
+    <PackageReference Include="Spectre.Console" Version="0.43.0" />
     <PackageReference Include="System.IO.Abstractions" Version="13.2.9" />
   </ItemGroup>
 </Project>
